Refactored code using resharper c++

+++ Space Shooter/ScientificConstants.hpp
@@ -1,8 +1,9 @@
 #ifndef scientificConstants_H
 #define scientificConstants_H
 
-namespace ScientificConstants {
-	inline constexpr double pi{ 3.14159265358979323846f };
+namespace ScientificConstants
+{
+	inline constexpr float pi{3.14159265358979323846f};
 }
 
 #endif
